Made the tests compileable again without widening the exception types on the signature.
Note that this means the the restart() method had to be pulled down.

A possible solution to this *may* have been to introduction a new class into the
inheritance hierarchy, but I would say that this would be unecessary indirection.

The real issue is the narrowness of the checked exceptions on the unit test
lifecycle methods.