drop toString in favour of constructing a "message" for the super class
By setting the message we get better error reporting in apps that call
getMessage() during error reporting rather than toString(). Clojure
appears to be an example of this.

And becasue the super class (RuntimeException) has a sensible
toString method that takes into account the class name and message we
do not our own toString method.