Added a test to ensure that if the client does not trust the server then it blows up. This is caught by setting an exceptionHandler, which inexplicably, isn't exposed through the ConnectionFactory (which is also the strangest Factory I've ever seen as it's stateful). So had to add API to the factory to allow the exceptionHandler to be set (which I *hate* because it'll be reused for every new connection which means if it's stateful then you're screwed, but the alternative is to multiply out the 3 overloaded newConnection methods to to 6 which I hate more).
And of course the test itself is racy, hence the synchronisation between the threads and the potential sitting there for 10 seconds doing nothing if it doesn't blow up as expected, and if the thread keeps being spuriously woken up then it'll continually wait.

However, it does indeed work for me.