correct exceptions for waitForConfirms
The two waitForConfirms can fail with exceptions in only two cases: 1)
when something happened to the underlying channel, we just throw
channel exception to the caller without wrapping it, and 2) when
waitForConfirmsOrDie received a nack, we throw an IOException("nacks
received").

The test is for the first case.  Since we don't have a way to cause a
nack, we can't test for the second case.