alternative test for data on a socket
Without this, ConnectionOpen.testCrazyProtocolHeader fails.  The test
connects, sends a crazy protocol header, reads the protocol version
from the server and checks that no more data is available on the
socket.

The changes on the broker seem to have tricked the java client into
thinking there's data available (in.available() returns 1).  Trying to
read said data fails, of course.  So, we now check that in.read()
fails.