Patch HeartbeatSender to work on Android when CPU sleeps
Executors.newSingleThreadScheduledExecutor does not wake up Android
if the CPU is asleep, so allow the client to pass in its own
heartbeat executor into the ConnectionFactory. On Android, the
client can construct an executor that wraps Android's AlarmManager.

Also, the optimization to skip sending heartbeats if there has been
recent activity was removed -- it doesn't work on Android because
System.nanoTime() doesn't include the time that the CPU is asleep.
On Android, real time can be determined with
SystemClock.elapsedRealtime() but this isn't cross-platform.

Also, the optimization to skip sending heartbeats after
recent activity was actually less efficient on Android because
it requires waking the CPU up more often. To maximize battery life,
it is better to wake up the CPU the absolute minimum necessary
even if it may require sending more heartbeat frames if there
has been recent activity.