Introduce default executor on ConnectionFactory
This makes it easier to share an executor between
connections.

Note that we do not have any default shared executor
because it would be impossible to know when/if it is
safe to shut it down. Even with reference counting
for connections using it in place, it is possible
that all connections produced by a factory
were closed only to create some new ones later.