Make SslContext creation more flexible
Introduce the SslContextFactory interface, in use in the ConnectionFactory
to create SslContext instance based on the connection name.
The existing ConnectionFactory#useSslProtocol() methods still work
the same way (they end using a SslContextFactory that returns the
same SslContext, whatever the name of the connection is).
This introduces a breaking change in the FrameHandlerFactory
by adding a new connectionName parameter. It should impact many
users, as this is more an internal API.

Fixes #241