Add option to ensure RPC reply is for the current request
During cluster node failovers some TimeoutExceptions during connection recovery
end up causing ClassCastExceptions because the reply for the timed out
RPC request comes in while a 2nd RPC request is waiting.

This commit add the ConnectionFactory#channelCheckRpcReplyType flag that, if enabled,
will make the channel check if a reply is compatible with the outstanding
RPC request. The flag is set to false by default but could be set to true in the future,
when the check logic is proven reliable.

This code was originally contributed by @vikinghawk.

Fixes #290