Get rid of read retry in NIO for plain connection
A FrameBuilder class has been introduced to be able to read a frame
in several network reads. The frame builder keeps track of
the partial frame state. This avoids pulling from the network
when the channel doesn't come back with any value from a network
read, which is against the NIO philosophy.

References #319