Load client version in a more reliable way
The code tries to read the version from a property file
(with a unique name) and falls back to Package#getImplementationVersion
in case of error. If Package#getImplementationVersion fails too,
a default hardcoded value is returned. Warns are logged for every failure.

This should make the retrieval more reliable in different contexts
(e.g. uber-JARs, OSGi).

References #334