Add RpcClient doCall/responseCall/primitiveCall overloads which include timeout
There will be use cases where a single shared timeout specified in the constructor
of the RpcClient will not match up with the expected timeout for different rpc calls
made using the rpc client. It's then convenient to be able to specifiy the timeout value
on a per method call basis. Otherwise the user of RpcClient would have to create
multiple instances of RpcClient, incurring the startup cost and having redundant
consumers of the response queue.