Trigger connection recovery on socket write error
Automatic connection recovery triggers now by default when a write
operation fails because of an IO exception. The recovery process
takes place in a dedicated thread so the write operation doesn't
wait (it receives the same IO exception immediatly).

The test to trigger the error has changed: it doesn't use manual
ack anymore, as this could sometimes block the broker and make
recovery fail (broker was busy re-enqueuing messages). The test
now sends a message in the consumer, which is enough to reproduce
the error.

Note the test against NIO is skipped right now, as it needs
additional care.

[#154263515]

References #341