Add timeout to work pool enqueuing
Making the work pool fail after it didn't manage  to enqueue
work for a given time makes the client more reactive
to broker overload. Note this usually happens to clients
that do not set QoS properly. Neverlethess, making the client
as early as possible can avoid hard-to-debug connection failure.

This complements the triggering of connection recovery on failed
write operations. Work pool enqueueing timeout is usefull for NIO,
where the same thread is used for both reading and writing (if
the thread is stuck waiting on work pool enqueueing, no write operation
can occur, and the TCP connection failure is never detected).

[#154263515]

Fixes #341