Reduce channel max to 2047
Max number of channels has been reduced from unlimited to 2048 as of
RabbitMQ 3.7.5. This is to avoid applications that leak channels to
starve the broker resources. 2048 has been picked up scientifically:
more than 1 K channels per connection is hard to justify, so we doubled
the value to leave existing applications some room.

Java client uses 2047 as it uses channel 0 for negotiation and error
handling.

Part of rabbitmq/rabbitmq-server#1593.

Fixes #366