Add test case for broken RpcClient on broker restart
The RpcClient registers a reply consumer on the direct-reply
pseudo-queue and doesn't cancel it  when the connection fails. The RpcClient
is then unusable by design (it nullifies the consumer on connection
failure). Nevertheless, the consumer must be cancelled after recovery to be
able to create a new RpcClient on the same channel.

Disabling topology recovery makes this work: the consumer isn't
re-registered during topology recovery and a new RpcClient can be
created on the same (recovered) channel.

References #382