Retry binding and consumer recovery
Auto-delete queues can go away during topology recovery, leading to
errors during binding and consumer recovery. A failed operation would
close the channel, making all the next operations fail as well.

This commit is an experimentation to see if recovering the channel,
re-creating the queue, and retrying the binding or consumer recovery is
a workable solution. It seems it is, even though the number of retries
can be significant (700 retries for 2000 queues).

A more general retry mechanism could be introduced to make topology
recovery more robust.