Support TLS hostname verification with HostnameVerifier
This is a solution for Java 6. For Java 7 and more, setting the server
validation algorithm on the SSLParameters is more appropriate.

This commit tries to make the activation of hostname verification as
simpler as possible for the developer. The client detects it's running
on Java 6 and sets up the hostname verifier from the Commons
HttpClient project (it's still possible to provide a customer
HostnameVerifier). The client uses the SSLParameters solution if
possible.

References #394