Fix and re-enable hanging test
The test could hang because a connection was in flow control and a queue
purge operation was stuck. The test now uses different connections to
publish and purge and use publisher confirms. Publisher confirms seems
necessary because the test was expecting the exact number of published
messages was purged from the queue. It's maybe expecting too much
considering publishing is asynchronous. Publisher confirms should make
the test more reliable.