Do not depend on generated classes for update-secret extension
Some CI testing jobs are using the 3.7 branch of codegen, which does not contain
the update-secret extension. The Java client then cannot be built
because it requires some generated code for the extension.

This commit adds the code to handle update-secret and thus
makes the generated code not necessary. This is just a workaround for
these testing jobs to succeed, it does not change the handling of the
update-secret extension in the client.

[#167029587]

References #626

(cherry picked from commit e3c662ea695126a0fc0c18b5640c5e3704eb3aff)