AutorecoveringConnection.maybeDeleteRecordedAutoDeleteExchange does not need to acquire a lock on this.consumers
It most likely was a copy-paste artefact introduced back in 2013-2014.
AutorecoveringConnection.maybeDeleteRecordedAutoDeleteQueue does need
to lock this.consumers as conditional queue deletion does need to
check the number of known consumers on that queue.

1aad565 addressed a potential deadlock caused by the unsafe
order of lock acquisitions. In #648 another similar issue
was discovered which #649 tried to address by acquiring a lock
on this.consumers early.

However, exchange cleanup does not need to lock this.consumers
as it does not mutate it.

Closes #648.