Change/remove usage of some deprecated JDK API
API deprecated in Java 16 or more. There are alternatives
(e.g. for certificate subject and issuer) for some, but not for
all (e.g. usage of the security manager to change thread, which
is no big deal as the changes to thread are minor and it is
likely nobody cares about such checks nowadays).

Fixes #709