Use count down latch in test
To avoid using Thread#stop(), which is deprecated since 1998.
The test does not explain why or if Thread#stop() is necessary
to the test. It's unlikely as the target thread has stopped already
when stop is called in the test.

(cherry picked from commit 6aaa203aa979307e7dc9326b7d9f0f077d8cc390)