Use publish confirms in visibility effect test
The test started to fail on a 2-node 3.10 cluster while
it was OK before. The test expected messages to be almost
immediately available in the queues after publishing, which
is not realistic, as publishing is asynchronous.

This commit enables publish confirms and waits for the confirms
to arrive before purging and checking the content of the queues,
which is more realistic in terms of expectations.

The bulk of the test remains asynchronous in case it would hang
like it used to do sometimes.