Handle process better in Host
The method to check if a command exists asks for the help
of this command now (instead of checking if the command name
is in the output of rabbitmqctl). This way we just have
to check the exit code (0 if the command exists).

We also use a timeout to wait for the end of the process.

A test hangs on the CI environment, these changes try to
fix this.

(cherry picked from commit 6b2251fe05684f39eca99eb846179df4c0d91e7e)