Increase max ConsumerWorkService block size to 256
As suggested in rabbitmq/rabbitmq-java-client#813. This has a few
minor effects on consumers:

 * A low single-digit % throughput gain
 * A comparable reduction in mean consumer latency

In general, it makes sense that consumers operating at peak throughput
should run operations in blocks close to the QoS prefetch used.
Since we usually recommend a value of 100-300 for environments that
focus on throughput, the new default of 256 makes sense.

The only negative effect I can think of a slightly higher GC pressure
which can increase variability of the aforementioned metrics.