Remove hard dependency on Jackson in OAuth 2 support
It's possible to override the token parsing method, to avoid
using Jackson, but the class would have a Jackson mapper
in its properties.

This commit introduces an abstraction to parse the token and
the default Jackson implementation is instantianted on the
fly.

(cherry picked from commit c0811fa13b02883166630d299bc48a21640cef55)

Conflicts:
	src/main/java/com/rabbitmq/client/impl/OAuth2ClientCredentialsGrantCredentialsProvider.java