Wrapping call in observe(...) and nulling sender
Due to the fact that in the tests we have both the producer and the consumer and they have the same parent, the zipkin graph looks bizarre.

With these changes we're changing the test sending code to simulate sending a message from a different service that will result in a creation of a new trace identifier. Due to this we will have 2 sets of trace ids created, one for sending and one for polling.

Sending: null_observation -> send -> receive message
Polling: test_span -> receive (very short)
(cherry picked from commit c854a0a7679136ef568526c72f9aa8fc2b3cf332)