Disable dead-lettering republishing test for 3.13+
A test that publishes a message with "x-death" header
already set and expects to see those appended in case
of another dead lettering.

After message container (3.13+), we avoid to mutate
messages, so dead letter events are appended to the header
if the message stays on the broker (e.g. is nacked, then expires).