Null-proof channel state retrieval in AbstractMetricsCollector
A channel message may have been closed by the time a metrics method is
called (e.g. when cancelling a consumer the metrics collector is called
after the RPC is completed), so the channel state may no longer be
there.

This commit null-proofs methods rely on the channel state retrieval.

Fixes #1592