Revert ConnectionFactory return type in some methods
Some changes from main snuck into 5.x during a merge. This affects 3
methods in ConnectionFactory that used to return void and returned
ConnectionFactory because of the unexpected merge.

The change is source-compatible but not binary-compatible. It should not
affect many users, but we try to stick to backward compatibility so we
go back to returning void.

References #1810